move test case and make it abstract
