+++ tests/Feature/Actions/CheckModuleIsInstalledTest.php
@@ -2,9 +2,6 @@
 
 use JustSteveKing\Laravel\ERP\Actions\CheckModuleIsInstalled;
 use JustSteveKing\Laravel\ERP\Models\Module;
-use JustSteveKing\Laravel\ERP\Tests\Feature\TestCase;
-
-uses(TestCase::class);
 
 it('returns a module if one is available with the given name')
     ->createModuleWithName('juststeveking/laravel-erp-crm')
