+++ tests/Feature/Commands/DiscoverModulesCommandTest.php
@@ -1,9 +1,6 @@
 <?php
 
 use JustSteveKing\Laravel\ERP\Models\Module;
-use JustSteveKing\Laravel\ERP\Tests\Feature\TestCase;
-
-uses(TestCase::class);
 
 it('can discover installed modules using the installed composer packages.')
     ->tap(fn() => runDiscoverCommand())
